re-architected subtitle/subtitle link sections
Previously, the subtitleLinks section was a child of the subtitle section. This made it difficult to place independently of the subtitle section for the mobile layout. Now, they are independent of each other. The alignment and justification of the items in the subtitle grid area changed a bit. Margins were added to separate the subtitle with the subtitle links