Fix header layout
Name, subtitle, and links now arranged in line with design.
May increase the distance between the top border and the name - will first test on phone.
Fixed error where all first a tags received special spacing. Now, this only applies to the
a tags in the nav elements. As a result, the elements in the heading (especially my name) are behaving more consistently.