Make small revisions
Mainly fixed margin stylings. Subtitle links have a smaller bottom margin.
Landing page shapes do not appear offscreen.
Reviewed elements on landing page, fixed it so that all widths are 100%vw. Trying
to find a solution for removing te grey bar that appears on the right side when the screen width is
<= 375px. So far no solution has been found, so I am still hiding any overflow.