Add Film-DB toprojects tab
Created a new image class - .image-container for the Film-DB screenshot. This sets object-fit to contain. It preserves the aspect ratio without zooming in/cropping the image. Works well for images that are not overly long and/or have dimensions similar to the height and width set for the image container